Split out admin pages, more instance/purchase fixes
