+++ tasks/dev.py
@@ -71,16 +71,16 @@ class MakeFakeTickets(Command):
 
                 full_count = random.choice([1] * 3 + [2, 3])
                 full_type = Product.get_by_name('general', 'full').get_cheapest()
-                full_tickets = Purchase.create_instances(user, full_type, 'GBP', full_count)
+                full_tickets = Purchase.create_purchases(user, full_type, 'GBP', full_count)
 
 
                 kids_count = random.choice([0] * 10 + [1, 2])
                 kids_type = Product.get_by_name('general', random.choice('u5', 'u16')).get_cheapest()
-                kids_tickets = Purchase.create_instances(user, kids_type, 'GBP', kids_count)
+                kids_tickets = Purchase.create_purchases(user, kids_type, 'GBP', kids_count)
 
                 vehicle_count = random.choice([0] * 2 + [1])
                 vehicle_type = Product.get_by_name('general', random.choice('parking', 'campervan')).get_cheapest()
-                vehicle_tickets = Purchase.create_instances(user, vehicle_type, 'GBP', vehicle_count)
+                vehicle_tickets = Purchase.create_purchases(user, vehicle_type, 'GBP', vehicle_count)
 
                 for t in full_tickets + kids_tickets + vehicle_tickets:
                     db.session.add(t)
