+++ templates/admin/_nav.html
@@ -30,7 +30,7 @@
         <ul class="dropdown-menu">
             {{ menuitem("Paid", ".tickets") }}
             {{ menuitem("Unpaid", ".tickets_unpaid") }}
-            {{ menuitem("Transfers", ".ticket_transfers") }}
+            {{ menuitem("Transfers", ".purchase_transfers") }}
             {{ menuitem("Free Tickets", ".tickets_choose_free") }}
             <li role="separator" class="divider"></li>
             {# menuitem("Product Groups", ".product_groups") #}
