+++ templates/admin/accounts/txn-suppress.html
