+++ templates/admin/payments/payments.html-allpayments
@@ -1,0 +1,57 @@
+{% extends "admin/base.html" %}
+{% set nav_active = 'admin_payments' %}
+
+{% macro render_toggle(field) -%}
+<label class="btn btn-primary {% if field.data -%} active {%- endif %}">{{ field() }}{{ field.label.text }}</label>
+{%- endmacro %}
+
+
+{% block body %}
+<h2>Payments</h2>
+
+<form action="" method="post">
+  <input name="csrf_token" type="hidden" value="{{ csrf_token() }}">
+  <span class="btn-group" data-toggle="buttons">
+    {{ render_toggle(form.paid) }}
+    {{ render_toggle(form.other) }}
+    {{ render_toggle(form.cancelled) }}
+  </span>
+
+  <span>
+    {{ form.update(class_="btn btn-primary") }}
+  </span>
+</form>
+
+<table class="table table-condensed table-striped">
+<thead><tr>
+  <th>Type</th>
+  <th>Name</th>
+  <th>Email</th>
+  <th>State</th>
+  <th>Tickets</th>
+  <th>Amount</th>
+  <th>Expires</th>
+  <th></th>
+</tr></thead>
+<tbody>
+{% for payment, first_expires, ticket_count in payments %}
+<tr>
+  <td><a href="{{ url_for('payment_invoice', payment_id=payment.id) }}">{{ payment.name }}</a></td>
+  <td>{{ payment.user.name }}</td>
+  <td>{{ payment.user.email }}</td>
+  <td>{{ payment.state }}</td>
+  <td>{{ ticket_count }}</td>
+  <td><b>{{ payment.amount | price(payment.currency) }}</b></td>
+  <td>{{ first_expires.strftime('%Y-%m-%d') }}</td>
+  <td>
+    {% if payment.provider == 'stripe' %}
+    <a class="btn btn-success" href="{{ url_for('admin_update_payment', payment_id=payment.id) }}">Update</a>
+    {% endif %}
+    <a class="btn btn-danger" href="{{ url_for('admin_cancel_payment', payment_id=payment.id) }}">Cancel</a>
+  </td>
+</tr>
+{% endfor %}
+</tbody>
+</table>
+
+{% endblock %}
