+++ templates/admin/products/purchase-transfers.html
@@ -1,5 +1,5 @@
 {% extends "admin/base.html" %}
-{% set nav_active = 'ticket_transfers' %}
+{% set nav_active = 'purchase_transfers' %}
 {% block body %}
 
 <h2>Ticket Transfers</h2>
